refactor(config,loop): split settings.py and domain_jobs.py under the module-health LOC cap
Both files were just under the 500-LOC cap on main; this branch's small
additions (triage_assessor config knobs + domain wiring) pushed them over.
Extract _LoopFlagAndCredentialSettings to config/settings_loop_flags.py and
the six optional single-scanner per-overlay job builders (+ their shared
_failed_e2e_scanner_for helper) to loop/domain_optional_scanner_jobs.py.
Update the three tests that patched the moved scanner factories by their
old teatree.loop.domain_jobs.* path.
