+++ tests/teatree_loop/test_pull_main_clone_scanner.py
@@ -676,12 +676,12 @@ class PullMainCloneScannerWiringTests(TestCase):
         fake_scanner = PullMainCloneScanner(repos=(("acme:acme-backend", Path("/x")),), cadence_hours=1)
         backend = OverlayBackends(name="acme", overlay=_FakeOverlay(["acme-backend"]))
         with (
-            patch("teatree.loop.domain_jobs._pull_main_clone_scanner_for", return_value=fake_scanner),
-            patch("teatree.loop.domain_jobs._architectural_review_scanner_for", return_value=None),
+            patch("teatree.loop.domain_optional_scanner_jobs._pull_main_clone_scanner_for", return_value=fake_scanner),
+            patch("teatree.loop.domain_optional_scanner_jobs._architectural_review_scanner_for", return_value=None),
             patch("teatree.loop.domain_jobs._pr_sweep_scanner_for", return_value=None),
             patch("teatree.loop.domain_jobs._codex_review_scanner_for", return_value=None),
             patch("teatree.loop.domain_jobs._slack_broadcasts_scanner_for", return_value=None),
-            patch("teatree.loop.domain_jobs._failed_e2e_scanner_for", return_value=None),
+            patch("teatree.loop.domain_optional_scanner_jobs._failed_e2e_scanner_for", return_value=None),
         ):
             jobs = _jobs_for_overlay_backend(backend)
