+++ tests/teatree_loop/test_task_sweep_wiring.py
@@ -163,12 +163,12 @@ class BuilderTests(TestCase):
         )()
         with (
             patch("teatree.loop.domain_jobs._task_sweep_scanner_for", return_value=fake),
-            patch("teatree.loop.domain_jobs._architectural_review_scanner_for", return_value=None),
+            patch("teatree.loop.domain_optional_scanner_jobs._architectural_review_scanner_for", return_value=None),
             patch("teatree.loop.domain_jobs._pr_sweep_scanner_for", return_value=None),
-            patch("teatree.loop.domain_jobs._pull_main_clone_scanner_for", return_value=None),
+            patch("teatree.loop.domain_optional_scanner_jobs._pull_main_clone_scanner_for", return_value=None),
             patch("teatree.loop.domain_jobs._codex_review_scanner_for", return_value=None),
             patch("teatree.loop.domain_jobs._slack_broadcasts_scanner_for", return_value=None),
-            patch("teatree.loop.domain_jobs._failed_e2e_scanner_for", return_value=None),
+            patch("teatree.loop.domain_optional_scanner_jobs._failed_e2e_scanner_for", return_value=None),
             patch("teatree.loop.domain_jobs._user_slack_id_for_overlay", return_value=""),
         ):
             jobs = _jobs_for_overlay_backend(backend)
